+++ c++/triqs/mesh/cyclat.hpp
@@ -37,15 +37,15 @@ namespace triqs::mesh {
      * @param bl_ bravais lattice
      * @param periodization_matrix such that $\tilde{a}_i = \sum_j N_{ij} a_j$
      */
-    cyclat(bravais_lattice const &bl_, matrix_const_view<int> periodization_matrix_) : bl(bl_), cluster_mesh{bl_.units(), periodization_matrix_} {}
+    cyclat(bravais_lattice const &bl_, matrix_const_view<long> periodization_matrix_) : bl(bl_), cluster_mesh{bl_.units(), periodization_matrix_} {}
 
     ///Construct  from three linear sizes assuming a cubic lattice (backward compatibility)
     cyclat(int L1 = 1, int L2 = 1, int L3 = 1)
-       : bl{eye<double>(3)}, cluster_mesh{eye<double>(3), matrix<int>{{{L1, 0, 0}, {0, L2, 0}, {0, 0, L3}}}} {}
+       : bl{eye<double>(3)}, cluster_mesh{eye<double>(3), matrix<long>{{{L1, 0, 0}, {0, L2, 0}, {0, 0, L3}}}} {}
 
     ///Construct from domain (bravais_lattice) and int L (linear size of Cluster mesh)
     cyclat(bravais_lattice const &bl_, int L)
-       : bl(bl_), cluster_mesh{bl_.units(), matrix<int>{{{L, 0, 0}, {0, bl_.ndim() >= 2 ? L : 1, 0}, {0, 0, bl_.ndim() >= 3 ? L : 1}}}} {}
+       : bl(bl_), cluster_mesh{bl_.units(), matrix<long>{{{L, 0, 0}, {0, bl_.ndim() >= 2 ? L : 1, 0}, {0, 0, bl_.ndim() >= 3 ? L : 1}}}} {}
 
     using domain_t = bravais_lattice;
     domain_t const &domain() const { return bl; }
